accessible/tests/mochitest/events/test_focus_listcontrols.xul

<?xml version="1.0"?>
<?xml-stylesheet href="chrome://global/skin" type="text/css"?>
<?xml-stylesheet href="chrome://mochikit/content/tests/SimpleTest/test.css"
                 type="text/css"?>

<window xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul"
        title="Accessible focus event testing">

  <script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js" />
  <script src="chrome://mochikit/content/tests/SimpleTest/EventUtils.js"/>

  <script type="application/javascript"
          src="../common.js" />
  <script type="application/javascript"
          src="../role.js" />
  <script type="application/javascript"
          src="../states.js" />
  <script type="application/javascript"
          src="../events.js" />

  <script type="application/javascript">
    //gA11yEventDumpID = "eventdump"; // debug stuff
    gA11yEventDumpToConsole = true; // debug stuff

    var gQueue = null;
    function doTests()
    {
      // Test focus events.
      gQueue = new eventQueue();

      gQueue.push(new synthFocus("richlistbox", new focusChecker("rlb_item1")));
      gQueue.push(new synthDownKey("rlb_item1", new focusChecker("rlb_item2")));
      gQueue.push(new synthFocus("multiselrichlistbox", new focusChecker("msrlb_item1")));
      gQueue.push(new synthDownKey("msrlb_item1", new focusChecker("msrlb_item2"), { shiftKey: true }));
      gQueue.push(new synthFocus("emptyrichlistbox", new focusChecker("emptyrichlistbox")));

      gQueue.push(new synthFocus("menulist"));
      gQueue.push(new synthClick("menulist", new focusChecker("ml_tangerine")));
      gQueue.push(new synthDownKey("ml_tangerine", new focusChecker("ml_marmalade")));
      gQueue.push(new synthEscapeKey("ml_marmalade", new focusChecker("menulist")));
if (!MAC) {
      // On Windows, items get selected during navigation.
      let expectedItem = WIN ? "ml_strawberry" : "ml_marmalade";
      gQueue.push(new synthDownKey("menulist", new nofocusChecker(expectedItem)));
      gQueue.push(new synthOpenComboboxKey("menulist", new focusChecker(expectedItem)));
      gQueue.push(new synthEnterKey(expectedItem, new focusChecker("menulist")));
} else {
      todo(false, "Bug 746531 - timeouts of last three menulist tests on OS X");
}

      // no focus events for unfocused list controls when current item is
      // changed.
      gQueue.push(new synthFocus("emptyrichlistbox"));

      gQueue.push(new changeCurrentItem("richlistbox", "rlb_item1"));
if (!MAC) {
      gQueue.push(new changeCurrentItem("menulist", WIN ? "ml_marmalade" : "ml_tangerine"));
}

      gQueue.invoke(); // Will call SimpleTest.finish();
    }

    SimpleTest.waitForExplicitFinish();
    addA11yLoadEvent(doTests);
  </script>

  <hbox flex="1" style="overflow: auto;">
    <body xmlns="http://www.w3.org/1999/xhtml">
      <a target="_blank"
         href="https://bugzilla.mozilla.org/show_bug.cgi?id=433418"
         title="Accessibles for focused HTML Select elements are not getting focused state">
        Mozilla Bug 433418
      </a>
      <a target="_blank"
         href="https://bugzilla.mozilla.org/show_bug.cgi?id=474893"
         title="List controls should fire a focus event on the selected child when tabbing or when the selected child changes while the list is focused">
        Mozilla Bug 474893
      </a>
      <a target="_blank"
         href="https://bugzilla.mozilla.org/show_bug.cgi?id=552368"
         title=" fire focus event on document accessible whenever the root or body element is focused">
        Mozilla Bug 552368
      </a>
      <p id="display"></p>
      <div id="content" style="display: none"></div>
      <pre id="test">
      </pre>
    </body>

    <vbox flex="1">
      <richlistbox id="richlistbox">
        <richlistitem id="rlb_item1">
          <description>A XUL Description!</description>
        </richlistitem>
        <richlistitem id="rlb_item2">
          <button label="A XUL Button"/>
        </richlistitem>
      </richlistbox>
      <richlistbox id="multiselrichlistbox" seltype="multiple">
        <richlistitem id="msrlb_item1">
          <description>A XUL Description!</description>
        </richlistitem>
        <richlistitem id="msrlb_item2">
          <button label="A XUL Button"/>
        </richlistitem>
      </richlistbox>
      <richlistbox id="emptyrichlistbox" seltype="multiple"/>

      <menulist id="menulist">
        <menupopup>
          <menuitem id="ml_tangerine" label="tangerine trees"/>
          <menuitem id="ml_marmalade" label="marmalade skies"/>
          <menuitem id="ml_strawberry" label="strawberry fields"/>
        </menupopup>
      </menulist>

      <vbox id="eventdump"/>
    </vbox>
  </hbox>
</window>
